+++ crates/driver/src/domain/competition/auction.rs
@@ -2,9 +2,10 @@ use {
     super::{order, Score},
     crate::{
         domain::{
-            competition::{self, auction, solution},
+            competition::{self, auction},
             eth,
             liquidity,
+            time,
         },
         infra::{self, blockchain, observe, Ethereum},
         util,
@@ -30,7 +31,7 @@ pub struct Auction {
     /// The tokens that are used in the orders of this auction.
     tokens: Tokens,
     gas_price: eth::GasPrice,
-    deadline: Deadline,
+    deadline: time::Deadline,
     score_cap: Score,
 }
 
@@ -39,7 +40,7 @@ impl Auction {
         id: Option<Id>,
         orders: Vec<competition::Order>,
         tokens: impl Iterator<Item = Token>,
-        deadline: Deadline,
+        deadline: time::Deadline,
         eth: &Ethereum,
         score_cap: Score,
     ) -> Result<Self, Error> {
@@ -104,7 +105,7 @@ impl Auction {
     }
 
     /// The deadline for the driver to start sending solution to autopilot.
-    pub fn deadline(&self) -> Deadline {
+    pub fn deadline(&self) -> time::Deadline {
         self.deadline
     }
 
@@ -395,40 +396,6 @@ impl From<Price> for eth::U256 {
 impl From<eth::U256> for Price {
     fn from(value: eth::U256) -> Self {
         Self(value.into())
-    }
-}
-
-/// Each auction has a deadline, limiting the maximum time that can be allocated
-/// to solving the auction.
-#[derive(Debug, Default, Clone, Copy)]
-pub struct Deadline(chrono::DateTime<chrono::Utc>);
-
-impl Deadline {
-    /// The remaining time left until the deadline, if any.
-    pub fn remaining(&self) -> Result<chrono::Duration, solution::DeadlineExceeded> {
-        let deadline = self.0 - infra::time::now();
-        if deadline < chrono::Duration::zero() {
-            Err(solution::DeadlineExceeded)
-        } else {
-            Ok(deadline)
-        }
-    }
-
-    #[must_use]
-    pub fn reduce(self, duration: chrono::Duration) -> Self {
-        Self(self.0 - duration)
-    }
-}
-
-impl From<chrono::DateTime<chrono::Utc>> for Deadline {
-    fn from(value: chrono::DateTime<chrono::Utc>) -> Self {
-        Self(value)
-    }
-}
-
-impl From<Deadline> for chrono::DateTime<chrono::Utc> {
-    fn from(value: Deadline) -> Self {
-        value.0
     }
 }
