+++ crates/driver/src/domain/mod.rs
@@ -3,6 +3,7 @@ pub mod eth;
 pub mod liquidity;
 pub mod mempools;
 pub mod quote;
+pub mod time;
 
 pub use {
     competition::Competition,
